refactor "new" components to regular
